LayoutTests/editing/pasteboard/ios/dom-paste-requires-user-gesture.html

<!DOCTYPE html> <!-- webkit-test-runner [ domPasteAllowed=false useFlexibleViewport=true ] -->
<html>
<meta name="viewport" content="width=device-width, initial-scale=1">
<head>
<script src="./resources/dom-paste-helper.js"></script>
<script src="../../../resources/js-test.js"></script>
<script src="../../../resources/ui-helper.js"></script>
<style>
body {
    margin: 0;
}

#copy {
    width: 100%;
    height: 50px;
    border: 1px dashed black;
}

#editor {
    width: 100%;
    height: 100px;
    border: 1px dashed silver;
    text-align: center;
}
</style>
</head>
<body>
<div id="editor" contenteditable></div>
<iframe id="copy" src="data:text/html,<div id='copy' style='font-size: 40px; text-align: center;'>Click here to copy</div>
    <script>
    copy.addEventListener('click', () => {
        getSelection().selectAllChildren(copy);
        document.execCommand('Copy');
        getSelection().removeAllRanges();
    });
    </script>"></iframe>
<div id="description"></div>
<div id="console"></div>
<script>
jsTestIsAsync = true;

const editor = document.getElementById("editor");

description("Verifies that no callout is shown when the page programmatically triggers paste outside the scope of user interaction. This test requires WebKitTestRunner.");

addEventListener("load", async () => {
    UIHelper.activateAt(160, 125).then(() => {
        editor.focus();
        waitForPasteMenu().then(finishJSTest);

        UIHelper.ensurePresentationUpdate().then(() => {
            internals.withUserGesture(() => shouldBe("document.execCommand('Paste')", "true"));
            shouldBe("document.execCommand('Paste')", "false");
            shouldBeEqualToString("editor.textContent", "Click here to copy");
        });
    });
});
</script>
</body>
</html>
